Replace generated build files with an autogen.sh script.
These files vary depending on the versions installed on
a developer's machine, so general practice is not to keep
them in version control, but to use a special script,

  ./autogen.sh

to set them up from installed packages. This just calls
the autoreconf command.

Not having them in the repo also speeds up checkouts, makes
for less noise in diffs and so on.